refactor(combos): isolate combo identifier helpers (split S11 L1/5)
